refactor: extract active input replay test harness
